Added selected property to RATreeNode, added showSelection, selectedCollapsedBorderColor, and selectedExpandedBorderColor properties to RATreeView. When set to YES, tapping a cell will add a border colour, by default it is a blue border if the cell is collapsed, a green border if the cell is expanded
Added expandOnlyIfSelected property to RATreeView. When set to YES, cells can only be expanded or collapsed if they are currently selected. It is recommended to also set showSelection to YES when using this option. Tapping a cell once will select it, tapping a second time will expand it, a further tap will collapse it. If already expanded before being tapped initially, the first tap will select the cell, a second tap will collapse it. In my personal experience, before implementing this, the tableview was very confusing and difficult to work with. I was using RATreeView with a hierarchy of configurations, each could contain sub configurations, and on the right side of the screen was a properties panel, showing the properties for the current selected config on the left in the tree view. I would find the treeView always expanding and collapsing when i didn't want it to, i was tapping on a cell to view the properties on the right side, not to expand or collapse it. Therefore adding only expand/collapse if selected fixed this issue, with first tap to select and view properties, a second tap would expand or collapse, and the border colours showed the state, blue for collapsed, green for expanded. This visually made it simpler to work with, knowing a cell must be green before pressing the plus button to add a child.