Interface Builder / Storyboard Support
The implementation is changed so the underlying UITableView can be configured from within Interface Builder, namely to make use of a Storyboard's advanced cell prototyping features. An example is included.

Along the way to make this possible, the internal RATableView is replaced by a NSProxy based solution of providing a scroll view to work with and of which the delegate can be set without breaking anything.